Made Robot.get(:id) mutable to allow easier manipulation of the exact instance, Show message when robot goes online/offline
Justification behind the Robot.get(:id) move is that the robot object could originate from anywhere, there were, and will be moving forward, points where we need to change robot information and save it back to the original location. As the infrastructure scales this will get more and more spread out thus more and more situations to account for. Passing my reference instead of by value allows us to bypass this, mutating the object directly.